requirements.txt
django>=4.2
torch>=2.0
torchvision>=0.15
timm>=0.9
Pillow>=9.0
